+++ packages/core/src/modules/staff/api/timesheets/time-projects/[id]/employees/route.ts
@@ -45,8 +45,8 @@ const rawBodySchema = z.object({}).passthrough()
 
 const listSchema = z
   .object({
-    page: z.coerce.number().min(1).default(1),
-    pageSize: z.coerce.number().min(1).max(100).default(50),
+    page: z.coerce.number().int().min(1).default(1),
+    pageSize: z.coerce.number().int().min(1).max(100).default(50),
     timeProjectId: z.string().uuid().optional(),
     status: z.string().optional(),
     sortField: z.string().optional(),
